X5 Bloque D: decimales en cantidad para productos que se venden por peso

+++ Kipu.Platform.Tests/InventoryItemOverflowTests.cs
@@ -6,14 +6,18 @@ namespace Kipu.Platform.Tests;
 ///     X4 A6: a pure unit test, no host and no database — InventoryItem.AddStock/
 ///     AdjustStock use `checked` arithmetic specifically so a StockUnit overflow
 ///     throws instead of silently wrapping to a negative value that reads as
-///     neither low nor out of stock (see StockRules.IsOutOfStock).
+///     neither low nor out of stock (see StockRules.IsOutOfStock). StockUnit
+///     became `decimal` in X5 Bloque D (fractional stock for products sold by
+///     weight), so the overflow boundary moved from int.MaxValue to
+///     decimal.MaxValue — still astronomically larger than any real bodega's
+///     stock, but the same `checked` guard still has to fire at its edge.
 /// </summary>
 public class InventoryItemOverflowTests
 {
     [Fact]
     public void AddStock_ThatWouldOverflowStockUnit_Throws()
     {
-        var item = new InventoryItem(productId: 1, warehouseId: 1, businessId: 1, stockUnit: int.MaxValue - 5,
+        var item = new InventoryItem(productId: 1, warehouseId: 1, businessId: 1, stockUnit: decimal.MaxValue - 5,
             minimumStock: 0);
 
         Assert.Throws<OverflowException>(() => item.AddStock(10));
@@ -22,7 +26,7 @@ public class InventoryItemOverflowTests
     [Fact]
     public void AdjustStock_ThatWouldOverflowStockUnit_Throws()
     {
-        var item = new InventoryItem(productId: 1, warehouseId: 1, businessId: 1, stockUnit: int.MaxValue - 5,
+        var item = new InventoryItem(productId: 1, warehouseId: 1, businessId: 1, stockUnit: decimal.MaxValue - 5,
             minimumStock: 0);
 
         Assert.Throws<OverflowException>(() => item.AdjustStock(10));
