+++ Kipu.Platform/Products/Domain/Model/Commands/UpdateMinimumStockCommand.cs
@@ -1,3 +1,3 @@
 namespace Kipu.Platform.Products.Domain.Model.Commands;
 
-public record UpdateMinimumStockCommand(int ProductId, int MinimumStock);
+public record UpdateMinimumStockCommand(int ProductId, decimal MinimumStock);
