+++ Kipu.Platform/Products/Interfaces/Rest/Resources/UpdateMinimumStockResource.cs
@@ -1,3 +1,3 @@
 namespace Kipu.Platform.Products.Interfaces.Rest.Resources;
 
-public record UpdateMinimumStockResource(int MinimumStock);
+public record UpdateMinimumStockResource(decimal MinimumStock);
