+++ Kipu.Platform/Sales/Domain/Model/Commands/SaleLineCommand.cs
@@ -1,3 +1,3 @@
 namespace Kipu.Platform.Sales.Domain.Model.Commands;
 
-public record SaleLineCommand(int ProductId, int Quantity);
+public record SaleLineCommand(int ProductId, decimal Quantity);
