+++ Kipu.Platform/Sales/Interfaces/Rest/Resources/SaleDetailResource.cs
@@ -1,3 +1,3 @@
 namespace Kipu.Platform.Sales.Interfaces.Rest.Resources;
 
-public record SaleDetailResource(int Id, int SaleId, int ProductId, int Quantity, decimal UnitPrice, decimal Discount, decimal Subtotal);
+public record SaleDetailResource(int Id, int SaleId, int ProductId, decimal Quantity, decimal UnitPrice, decimal Discount, decimal Subtotal);
